feat(projects): browser-direct writes slice B (create/update/duplicate/saveAsTemplate/delete)
Converts the last project write server actions to browser-direct calls into
the existing createNative/updateNative/duplicateNative/saveAsTemplateNative/
deleteNative Convex mutations, via a new set of useProjectWrites(orgId)
methods alongside Slice A's archive/deleteTemplate. src/server/projects.ts
is now reads-only.

Hardens the two mutations for their new browser-direct exposure:
duplicateNative/saveAsTemplateNative now resolve orgDefaultTaxRate
in-mutation from the orgSettings mirror (extracted resolveOrgDefaultTaxRate
into shared convex/lib/orgSettings.ts) instead of trusting a client-supplied
value, and updateNative/createNative validate money-adjacent fields
(taxRate/discountPercent/depositPercent/depositPaid/invoicedTotal) via a new
assertProjectMoneyFields guard, mirroring the existing line-item money guards.
