+++ convex/lib/orgSettings.ts
@@ -1,0 +1,12 @@
+import type { MutationCtx, QueryCtx } from "../_generated/server";
+
+/** Org default tax rate from the Convex orgSettings mirror (source of truth; the
+ *  Postgres column is deprecated). null when unset. Resolved IN-mutation so browser
+ *  callers can't spoof a money-affecting tax rate. */
+export async function resolveOrgDefaultTaxRate(ctx: MutationCtx | QueryCtx, orgId: string): Promise<number | null> {
+  const row = await ctx.db
+    .query("orgSettings")
+    .withIndex("by_organizationId", (q) => q.eq("organizationId", orgId))
+    .first();
+  return row?.defaultTaxRate ?? null;
+}
